pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.3.1",
    "dspy>=3.0.4",
    "fastapi>=0.126.0",
    "fastmcp>=2.14.1",
    "gepa>=0.0.17",
    "langchain>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "mlflow==3.8.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "jupyterlab>=4.5.1",
    "ruff>=0.14.10",
    "ty>=0.0.7",
]

[tool.typy]
# 忽略某些規則以減少誤報
disable = [
    "invalid-argument-type",  # StateGraph 和其他第三方庫的類型推斷問題
    "unknown-argument",       # ChatOpenAI 等庫的參數檢查
    "possibly-missing-attribute",  # mlflow.genai 等動態屬性
]
